+++ src/index.js
@@ -1,18 +1,10 @@
 const serverless = require("serverless-http");
 const express = require("express");
-const app = express();
-const { handleWebhookEvent } = require("./handlers/webhookHandler");
+const { lineWebhookHandler } = require("./handlers/lineWebhookHandler");
 const lineConfig = require("../config/lineConfig");
+const app = express();
 
 app.use("/webhook", require("@line/bot-sdk").middleware(lineConfig));
-app.post("/webhook", (req, res) => {
-  Promise.all(req.body.events.map(handleWebhookEvent))
-    .then((result) => res.json(result))
-    .catch((err) => {
-      console.error(err);
-      res.status(500).end();
-    });
-});
+app.post("/webhook", lineWebhookHandler);
 
-const server = serverless(app);
-module.exports.handler = server;
+module.exports.handler = serverless(app);
